add send points function

+++ lib/features/profile/models/send_hearts_request_model.dart
@@ -1,5 +1,5 @@
 class SendHeartsRequestModel {
-  int? hearts;
+  double? hearts;
   String? phoneNumber;
 
   SendHeartsRequestModel({this.hearts, this.phoneNumber});
